// src/types.d.ts
export interface Project {
  /** A unique ID number that identifies a project. */
  id: string;

  /** A project's unique slug – part of the project's URL based on its name. */
  slug: string;

  /**  */
  title: string;
  /** Optional description of project. */
  description?: string;
  /**  */
  image?: ImageMetadata | string;

  /**  */
  tags?: Taxonomy[];

  /** Whether this project is featured */
  featured: boolean;
  /** Whether this project has a case study */
  caseStudy: boolean;

  /** Link to live project */
  link?: string;
  /** Link to GitHub repository */
  github?: string;

  /**  */
  publishDate: Date;

  /**  */
  Content?: AstroComponentFactory;
  content?: string;
}

export interface Taxonomy {
  slug: string;
  title: string;
}
// ...
export interface Widget {
  id?: string;
  isDark?: boolean;
  bg?: string;
  classes?: Record<string, string | Record<string, string>>;
}

export interface Headline {
  title?: string;
  subtitle?: string;
  tagline?: string;
  classes?: Record<string, string>;
}

interface TeamMember {
  name?: string;
  job?: string;
  image?: Image;
  socials?: Array<Social>;
  description?: string;
  classes?: Record<string, string>;
}

interface Social {
  icon?: string;
  href?: string;
}

export interface Stat {
  amount?: number | string;
  title?: string;
  icon?: string;
}

export interface Item {
  title?: string;
  description?: string;
  icon?: string;
  classes?: Record<string, string>;
  callToAction?: CallToAction;
  image?: Image;
}
// ...
export interface CallToAction extends Omit<HTMLAttributes<'a'>, 'slot'> {
  variant?: 'primary' | 'secondary' | 'tertiary' | 'link';
  text?: string;
  icon?: string;
  classes?: Record<string, string>;
  type?: 'button' | 'submit' | 'reset';
}
// ...
export interface Content extends Omit<Headline, 'classes'>, Widget {
  content?: string;
  image?: string | unknown;
  items?: Array<Item>;
  columns?: number;
  isReversed?: boolean;
  isAfterContent?: boolean;
  callToAction?: CallToAction;
}